refactor(sample-react): rename callback stories to get-value
