+++ packages/samples/react/src/components/input-checkbox/get-value.tsx
