+++ packages/samples/react/src/components/input-checkbox/routes.ts
@@ -1,13 +1,13 @@
 import type { Routes } from '../../shares/types';
 import { InputCheckboxBasic } from './basic';
 import { InputCheckboxButton } from './button';
-import { InputCheckboxOnInputOnChange } from './on-input-on-change';
+import { InputCheckboxOnInputOnChange } from './get-value';
 import { InputCheckboxSwitch } from './switch';
 
 export const INPUT_CHECKBOX_ROUTES: Routes = {
 	'input-checkbox': {
-		'on-input-on-change': InputCheckboxOnInputOnChange,
 		basic: InputCheckboxBasic,
+		'get-value': InputCheckboxOnInputOnChange,
 		switch: InputCheckboxSwitch,
 		button: InputCheckboxButton,
 	},
