+++ packages/samples/react/src/components/input-color/get-value.tsx
