+++ packages/samples/react/src/components/input-color/routes.ts
@@ -1,10 +1,10 @@
 import type { Routes } from '../../shares/types';
 import { InputColorBasic } from './basic';
-import { InputColorOnInputOnChange } from './on-input-on-change';
+import { InputColorOnInputOnChange } from './get-value';
 
 export const INPUT_COLOR_ROUTES: Routes = {
 	'input-color': {
-		'on-input-on-change': InputColorOnInputOnChange,
 		basic: InputColorBasic,
+		'get-value': InputColorOnInputOnChange,
 	},
 };
