+++ packages/samples/react/src/components/input-date/get-value.tsx
