+++ packages/samples/react/src/components/input-email/get-value.tsx
