+++ packages/samples/react/src/components/input-email/routes.ts
@@ -1,10 +1,10 @@
 import type { Routes } from '../../shares/types';
 import { InputEmailBasic } from './basic';
-import { InputEmailOnInputOnChange } from './on-input-on-change';
+import { InputEmailOnInputOnChange } from './get-value';
 
 export const INPUT_EMAIL_ROUTES: Routes = {
 	'input-email': {
-		'on-input-on-change': InputEmailOnInputOnChange,
 		basic: InputEmailBasic,
+		'get-value': InputEmailOnInputOnChange,
 	},
 };
