+++ packages/samples/react/src/components/input-file/get-value.tsx
