+++ packages/samples/react/src/components/input-file/routes.ts
@@ -1,10 +1,10 @@
 import type { Routes } from '../../shares/types';
 import { InputFileBasic } from './basic';
-import { InputFileOnInputOnChange } from './on-input-on-change';
+import { InputFileOnInputOnChange } from './get-value';
 
 export const INPUT_FILE_ROUTES: Routes = {
 	'input-file': {
-		'on-input-on-change': InputFileOnInputOnChange,
 		basic: InputFileBasic,
+		'get-value': InputFileOnInputOnChange,
 	},
 };
