+++ packages/samples/react/src/components/input-number/get-value.tsx
