+++ packages/samples/react/src/components/input-number/routes.ts
@@ -1,12 +1,12 @@
 import type { Routes } from '../../shares/types';
 import { InputNumberBasic } from './basic';
+import { InputNumberOnInputOnChange } from './get-value';
 import { InputNumberNumberFormatter } from './number-formatter';
-import { InputNumberOnInputOnChange } from './on-input-on-change';
 
 export const INPUT_NUMBER_ROUTES: Routes = {
 	'input-number': {
-		'on-input-on-change': InputNumberOnInputOnChange,
 		basic: InputNumberBasic,
+		'get-value': InputNumberOnInputOnChange,
 		'number-formatter': InputNumberNumberFormatter,
 	},
 };
