+++ packages/samples/react/src/components/input-password/get-value.tsx
