+++ packages/samples/react/src/components/input-password/routes.ts
@@ -1,12 +1,12 @@
 import type { Routes } from '../../shares/types';
 import { InputPasswordBasic } from './basic';
-import { InputPasswordOnInputOnChange } from './on-input-on-change';
+import { InputPasswordOnInputOnChange } from './get-value';
 import { InputPasswordShowPassword } from './show-password';
 
 export const INPUT_PASSWORD_ROUTES: Routes = {
 	'input-password': {
-		'on-input-on-change': InputPasswordOnInputOnChange,
 		basic: InputPasswordBasic,
+		'get-value': InputPasswordOnInputOnChange,
 		'show-password': InputPasswordShowPassword,
 	},
 };
