+++ packages/samples/react/src/components/input-radio/get-value.tsx
