+++ packages/samples/react/src/components/input-radio/routes.ts
@@ -1,13 +1,13 @@
 import type { Routes } from '../../shares/types';
 import { InputRadioBasic } from './basic';
+import { InputRadioOnInputOnChange } from './get-value';
 import { InputRadioHorizontal } from './horizontal';
 import { InputRadioObjectValue } from './objectValue';
-import { InputRadioOnInputOnChange } from './on-input-on-change';
 
 export const INPUT_RADIO_ROUTES: Routes = {
 	'input-radio': {
-		'on-input-on-change': InputRadioOnInputOnChange,
 		basic: InputRadioBasic,
+		'get-value': InputRadioOnInputOnChange,
 		horizontal: InputRadioHorizontal,
 		object: InputRadioObjectValue,
 	},
