+++ packages/samples/react/src/components/input-range/get-value.tsx
