+++ packages/samples/react/src/components/input-range/routes.ts
@@ -1,10 +1,10 @@
 import type { Routes } from '../../shares/types';
 import { InputRangeBasic } from './basic';
-import { InputRangeOnInputOnChange } from './on-input-on-change';
+import { InputRangeOnInputOnChange } from './get-value';
 
 export const INPUT_RANGE_ROUTES: Routes = {
 	'input-range': {
-		'on-input-on-change': InputRangeOnInputOnChange,
 		basic: InputRangeBasic,
+		'get-value': InputRangeOnInputOnChange,
 	},
 };
