+++ packages/samples/react/src/components/input-text/get-value.tsx
