+++ packages/samples/react/src/components/input-text/routes.ts
@@ -5,10 +5,10 @@ import { InputTextBasic } from './basic';
 import { InputTextCounter } from './counter';
 import { InputTextDisabled } from './disabled';
 import { InputTextExpertSlot } from './expert-slot';
+import { InputTextOnInputOnChange } from './get-value';
 import { InputTextHideLabel } from './hide-label';
 import { InputTextHideMsg } from './hide-msg';
 import { InputTextMessageTypes } from './message-types';
-import { InputTextOnInputOnChange } from './on-input-on-change';
 import { InputTextPattern } from './pattern';
 import { InputTextPlaceholder } from './placeholder';
 import { InputTextReadonly } from './readonly';
@@ -18,8 +18,8 @@ import { InputTextFormatterDemo } from './text-formatter';
 
 export const INPUT_TEXT_ROUTES: Routes = {
 	'input-text': {
-		'on-input-on-change': InputTextOnInputOnChange,
 		basic: InputTextBasic,
+		'get-value': InputTextOnInputOnChange,
 		'message-types': InputTextMessageTypes,
 		placeholder: InputTextPlaceholder,
 		disabled: InputTextDisabled,
