+++ packages/samples/react/src/components/select/get-value.tsx
