+++ packages/samples/react/src/components/select/routes.ts
@@ -1,10 +1,10 @@
 import type { Routes } from '../../shares/types';
 import { SelectBasic } from './basic';
-import { SelectOnInputOnChange } from './on-input-on-change';
+import { SelectOnInputOnChange } from './get-value';
 
 export const SELECT_ROUTES: Routes = {
 	select: {
-		'on-input-on-change': SelectOnInputOnChange,
 		basic: SelectBasic,
+		'get-value': SelectOnInputOnChange,
 	},
 };
