+++ packages/samples/react/src/components/single-select/get-value.tsx
