+++ packages/samples/react/src/components/textarea/get-value.tsx
